Add Polymorphism

+++ src/oop/inheritance/SavingsAccount.java
@@ -1,9 +1,10 @@
 package oop.inheritance;
 
 public class SavingsAccount extends BankAccount {
+
     private int transactions;
 
-    public SavingsAccount(int bankCode, String accountNumber, double balance, int transactions) {
+    public SavingsAccount(int bankCode, String accountNumber, int balance, int transactions) {
         super(bankCode, accountNumber, balance);
         this.transactions = transactions;
     }
@@ -15,5 +16,4 @@ public class SavingsAccount extends BankAccount {
     public void setTransactions(int transactions) {
         this.transactions = transactions;
     }
-
 }
